fix(PoliciesTable): Use RBAC hook correctly
