+++ src/PresentationalComponents/PoliciesTable/PoliciesTable.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import propTypes from 'prop-types';
-import { withRouter } from 'react-router-dom';
 import { Button } from '@patternfly/react-core';
 import { COMPLIANCE_TABLE_DEFAULTS } from '@/constants';
 import { BackgroundLink } from 'PresentationalComponents';
@@ -9,7 +8,7 @@ import useFeature from 'Utilities/hooks/useFeature';
 import columns, { exportableColumns } from './Columns';
 import * as Filters from './Filters';
 import { emptyRows } from '../../Utilities/hooks/useTableTools/Components/NoResultsTable';
-import { usePermissions } from '@redhat-cloud-services/frontend-components-utilities/RBACHook';
+import useActionResolver from './hooks/useActionResolvers';
 
 const DedicatedAction = () => (
   <BackgroundLink
@@ -22,37 +21,10 @@ const DedicatedAction = () => (
   </BackgroundLink>
 );
 
-export const PoliciesTable = ({ policies, location, history }) => {
+export const PoliciesTable = ({ policies }) => {
   const manageColumnsEnabled = useFeature('manageColumns');
   const filters = Object.values(Filters);
-  const { hasAccess: hasDeleteAccess, isLoading: isDeleteAccessLoading } =
-    usePermissions('compliance', `/scappolicies/XYZ/delete`, false, true);
-  const { hasAccess: hasEditAccess, isLoading: isEditAccessLoading } =
-    usePermissions('compliance', `/scappolicies/XYZ/edit`, false, true);
-
-  const onClick = (to, { itemId: policyId }) => {
-    const policy = policies.find((policy) => policy.id === policyId);
-    history.push(`/scappolicies/${policyId}/delete`, {
-      policy,
-      background: location,
-      state: { policy },
-    });
-  };
-
-  const actionResolver = () => [
-    {
-      title: 'Delete policy',
-      isDisabled: !isDeleteAccessLoading && !hasDeleteAccess,
-      onClick: (_event, _index, policy) =>
-        onClick(`/scappolicies/${policy.itemId}/delete`, policy),
-    },
-    {
-      title: 'Edit policy',
-      isDisabled: !isEditAccessLoading && !hasEditAccess,
-      onClick: (_event, _index, policy) =>
-        onClick(`/scappolicies/${policy.itemId}/edit`, policy),
-    },
-  ];
+  const actionResolver = useActionResolver(policies);
 
   return (
     <TableToolsTable
@@ -82,12 +54,10 @@ export const PoliciesTable = ({ policies, location, history }) => {
 
 PoliciesTable.propTypes = {
   policies: propTypes.array.isRequired,
-  history: propTypes.object.isRequired,
-  location: propTypes.object.isRequired,
 };
 
 PoliciesTable.defaultProps = {
   policies: [],
 };
 
-export default withRouter(PoliciesTable);
+export default PoliciesTable;
